requirements.txt
git+https://github.com/pknull/rpg-dice.git@master
git+https://github.com/pknull/rpg-card.git@master
git+https://github.com/pknull/rpg-flip.git@master
discord.py
dotenv
requests
openai
